Separate out ol.style.IconImage
